+++ testdata/cheats/Mapping.t.sol
@@ -5,15 +5,15 @@ import "ds-test/test.sol";
 import "./Cheats.sol";
 
 contract RecordMapping {
-    int length;
-    mapping(address => int) data;
-    mapping(int => mapping(int => int)) nestedData;
+    int256 length;
+    mapping(address => int256) data;
+    mapping(int256 => mapping(int256 => int256)) nestedData;
 
-    function setData(address addr, int value) public {
+    function setData(address addr, int256 value) public {
         data[addr] = value;
     }
 
-    function setNestedData(int i, int j) public {
+    function setNestedData(int256 i, int256 j) public {
         nestedData[i][j] = i * j;
     }
 }
@@ -35,34 +35,38 @@ contract RecordMappingTest is DSTest {
         bytes32 key;
         bytes32 parent;
 
-        bytes32 dataSlot = bytes32(uint(1));
-        bytes32 nestDataSlot = bytes32(uint(2));
-        assertEq(uint(cheats.getMappingLength(address(target), dataSlot)), 1, "number of data is incorrect");
-        assertEq(uint(cheats.getMappingLength(address(this), dataSlot)), 0, "number of data is incorrect");
-        assertEq(uint(cheats.getMappingLength(address(target), nestDataSlot)), 2, "number of nestedData is incorrect");
+        bytes32 dataSlot = bytes32(uint256(1));
+        bytes32 nestDataSlot = bytes32(uint256(2));
+        assertEq(uint256(cheats.getMappingLength(address(target), dataSlot)), 1, "number of data is incorrect");
+        assertEq(uint256(cheats.getMappingLength(address(this), dataSlot)), 0, "number of data is incorrect");
+        assertEq(
+            uint256(cheats.getMappingLength(address(target), nestDataSlot)), 2, "number of nestedData is incorrect"
+        );
 
         bytes32 dataValueSlot = cheats.getMappingSlotAt(address(target), dataSlot, 0);
         (found, key, parent) = cheats.getMappingKeyAndParentOf(address(target), dataValueSlot);
         assert(found);
         assertEq(address(uint160(uint256(key))), address(this), "key of data[i] is incorrect");
         assertEq(parent, dataSlot, "parent of data[i] is incorrect");
-        assertGt(uint(dataValueSlot), 0);
-        assertEq(uint(cheats.load(address(target), dataValueSlot)), 100);
+        assertGt(uint256(dataValueSlot), 0);
+        assertEq(uint256(cheats.load(address(target), dataValueSlot)), 100);
 
-        for (uint k; k < cheats.getMappingLength(address(target), nestDataSlot); k++) {
+        for (uint256 k; k < cheats.getMappingLength(address(target), nestDataSlot); k++) {
             bytes32 subSlot = cheats.getMappingSlotAt(address(target), nestDataSlot, k);
             (found, key, parent) = cheats.getMappingKeyAndParentOf(address(target), subSlot);
-            uint i = uint(key);
+            uint256 i = uint256(key);
             assertEq(parent, nestDataSlot, "parent of nestedData[i][j] is incorrect");
-            assertEq(uint(cheats.getMappingLength(address(target), subSlot)), 1, "number of nestedData[i] is incorrect");
+            assertEq(
+                uint256(cheats.getMappingLength(address(target), subSlot)), 1, "number of nestedData[i] is incorrect"
+            );
             bytes32 leafSlot = cheats.getMappingSlotAt(address(target), subSlot, 0);
             (found, key, parent) = cheats.getMappingKeyAndParentOf(address(target), leafSlot);
-            uint j = uint(key);
+            uint256 j = uint256(key);
             assertEq(parent, subSlot, "parent of nestedData[i][j] is incorrect");
             assertEq(j, 10);
-            assertEq(uint(cheats.load(address(target), leafSlot)), i * j, "value of nestedData[i][j] is incorrect");
+            assertEq(uint256(cheats.load(address(target), leafSlot)), i * j, "value of nestedData[i][j] is incorrect");
         }
         cheats.stopMappingRecording();
-        assertEq(uint(cheats.getMappingLength(address(target), dataSlot)), 0, "number of data is incorrect");
+        assertEq(uint256(cheats.getMappingLength(address(target), dataSlot)), 0, "number of data is incorrect");
     }
 }
